Consider sequence position when picking colors.
This commit was motivated by the desire to color the "wildtype" symbol
in each stack differently, so I could easily see whether or not each
position was mutated.

To do this, I had to generalize the color scheme framework a bit.  Color
schemes are now collections of color rules, which are objects that can
return a color given a sequence index, a symbol, and the ranked order of
that symbol.  I then turned ColorGroup into the SymbolColor rule and
added IndexColor and RefSeqColor rules.

I also had to generalize the *.eps template to support the more flexible
color schemes.  I added a color argument to ShowSymbol and used the
color scheme to generate this argument for each symbol.  Then I removed
color_dict and SetColor, which were both no longer necessary.  The new
*.eps template is actually simpler and more powerful than the old one.