Allowed TRANSFAC PWMs containing any subset of a provided alphabet to be validly parsed.
- modified alphabet verification to check that the de facto alphabet is a subset of that provided
- refactored, to not attempt to reindex the PWM to any other non–de facto alphabet

It is not clear what the correct form of reindexing ought be in such contexts.
Accordingly, we simply use the de facto alphabet,
provided that it is not a superset of a user-defined alphabet.

Custom PWM alphabets can now be automatically parsed, without the specification of
an `--alphabet` parameter, but `--color` parameters should still be provided,
when colouring of custom alphabets is desired.